# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "paths": {
    "/api/health": {
      "get": {
        "summary": "Health",
        "operationId": "health_api_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResult"
                }
              }
            }
          }
        }
      }
    },
    "/api/posts": {
      "get": {
        "summary": "Get Posts",
        "operationId": "get_posts_api_posts_get",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 10,
              "title": "Page Size"
            }
          },
          {
            "name": "nature",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/PostNatureEnum"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Nature"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedRecords_PostPublic_"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized access",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Invalid or missing data provided",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create Post",
        "operationId": "create_post_api_posts_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePostInputDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostPublic"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized access",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Invalid or missing data provided",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/posts/{id}": {
      "get": {
        "summary": "Get Post",
        "operationId": "get_post_api_posts__id__get",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostPublic"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized access",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Invalid or missing data provided",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "summary": "Update Post",
        "operationId": "update_post_api_posts__id__patch",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePostInputDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostPublic"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized access",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Invalid or missing data provided",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete Post",
        "operationId": "delete_post_api_posts__id__delete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "description": "Unauthorized access",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Invalid or missing data provided",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "CreatePostInputDto": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Title"
          },
          "content": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Content"
          },
          "nature": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PostNatureEnum",
                "maxLength": 255
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "title",
          "content"
        ],
        "title": "CreatePostInputDto"
      },
      "HTTPErrorResponse": {
        "properties": {
          "detail": {
            "type": "string",
            "title": "Detail",
            "examples": [
              "error message"
            ]
          }
        },
        "type": "object",
        "required": [
          "detail"
        ],
        "title": "HTTPErrorResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail",
            "examples": [
              [
                {
                  "input": {},
                  "loc": [
                    "body",
                    "title"
                  ],
                  "msg": "Field required",
                  "type": "missing"
                }
              ]
            ]
          }
        },
        "type": "object",
        "required": [
          "detail"
        ],
        "title": "HTTPValidationError"
      },
      "HealthResult": {
        "properties": {
          "api": {
            "type": "boolean",
            "title": "Api"
          },
          "database": {
            "type": "boolean",
            "title": "Database"
          }
        },
        "type": "object",
        "required": [
          "api",
          "database"
        ],
        "title": "HealthResult"
      },
      "PaginatedRecords_PostPublic_": {
        "properties": {
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size"
          },
          "total_count": {
            "type": "integer",
            "title": "Total Count"
          },
          "total_pages": {
            "type": "integer",
            "title": "Total Pages"
          },
          "records": {
            "items": {
              "$ref": "#/components/schemas/PostPublic"
            },
            "type": "array",
            "title": "Records"
          }
        },
        "type": "object",
        "required": [
          "page",
          "page_size",
          "total_count",
          "total_pages",
          "records"
        ],
        "title": "PaginatedRecords[PostPublic]"
      },
      "PostNatureEnum": {
        "type": "string",
        "enum": [
          "scientific",
          "standard",
          "journal"
        ],
        "title": "PostNatureEnum"
      },
      "PostPublic": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Title"
          },
          "content": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Content"
          },
          "nature": {
            "$ref": "#/components/schemas/PostNatureEnum",
            "maxLength": 255,
            "minLength": 1
          },
          "user_uid": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "User Uid"
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "title",
          "content",
          "nature",
          "user_uid",
          "id",
          "created_at",
          "updated_at"
        ],
        "title": "PostPublic"
      },
      "UpdatePostInputDto": {
        "properties": {
          "title": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "content": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Content"
          },
          "nature": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PostNatureEnum",
                "maxLength": 255
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "title": "UpdatePostInputDto"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Loc"
          },
          "msg": {
            "type": "string",
            "title": "Msg"
          },
          "type": {
            "type": "string",
            "title": "Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "title": "Ctx"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type",
          "input",
          "ctx"
        ],
        "title": "ValidationError"
      }
    }
  }
}